edge-node-components/requirements.txt
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
docopt==0.6.2
greenlet==2.0.2
idna==3.4
inflect==6.0.4
install==1.3.5
mariadb==1.1.6
numpy~=1.24.3
packaging==23.0
pipreqs==0.4.11
pydantic==1.10.7
pyproject_hooks==1.0.0
python-dotenv==1.0.0
requests~=2.29.0
sqlacodegen @ git+https://github.com/agronholm/sqlacodegen@31583dace790f4b69b98b31576ac6ddd99155ee9
SQLAlchemy~=2.0.15
typing_extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9

jsonschema~=4.17.3
paho-mqtt~=1.6.1
pandas~=2.0.1
pythermalcomfort~=2.8.0
alembic~=1.11.1
pytest~=7.4.0
pytz~=2023.3